refactor: add general search page + fix: search path correction
