+++ src/Pages/MyListPage.jsx
@@ -3,6 +3,7 @@ import Footer from "../Components/Footer";
 import MainTopBar from "../Components/MainPage/MainTopBar";
 import MyListBody from "../Components/MyListPage/MyListBody";
 import { makeStyles } from '@material-ui/core/styles';
+import SearchBody from "../Components/Search/SearchBody";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -15,11 +16,13 @@ const useStyles = makeStyles((theme) => ({
 function MyListPage() {
   const classes = useStyles();
   const listFromRedux = useSelector((state) => state.NetflixReducer.myList);
+  const globalSearch = useSelector((state) => state.NetflixReducer.globalSearch);
+
 
   return (
     <div className={classes.root}>
       <MainTopBar />
-      <MyListBody myList={ listFromRedux }/>
+      {globalSearch ? <SearchBody /> : <MyListBody myList={ listFromRedux }/>}
       <Footer />
     </div>
   )
